+++ lib/src/month_picker_widgets/header/header.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '/src/helpers/controller.dart';
-import '/src/month_picker_widgets/header/header_row.dart';
-import '/src/month_picker_widgets/header/header_selected_date.dart';
+import '/month_picker_dialog.dart';
 
 class PickerHeader extends StatelessWidget {
   const PickerHeader({