+++ lib/src/month_selector/month_year_grid.dart
@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
-import '/src/helpers/controller.dart';
-import '/src/helpers/locale_utils.dart';
-
-import 'month_button.dart';
+import '/month_picker_dialog.dart';
 
 class MonthYearGridBuilder extends StatelessWidget {
   const MonthYearGridBuilder({